Move get_dummy_pr_with_list_of_files so other files can use it
